% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins_to_bedgraph.R
\name{bins_to_bedgraph}
\alias{bins_to_bedgraph}
\title{Bins to bed graph.}
\usage{
bins_to_bedgraph(bin_df, min_value = 3, filename = "test.bed")
}
\arguments{
\item{bin_df}{Data frame with bins.}

\item{min_value}{Minimum value to retain the bin.}

\item{filename}{Path to a local file on drive to save the resulting file.}
}
\value{
full path to the file that was written
}
\description{
This function will generate bed file.
}
