% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_output.R
\name{cache_output}
\alias{cache_output}
\title{Cache output.}
\usage{
cache_output(
  result_df,
  function_name,
  clobber_mode = F,
  get_existing = F,
  function_params = list(region = "chr3:98300000-198022430", bin_size = 2000, seq_type =
    "genome"),
  additional_details = list(foreground = "DLBCL_FL_BL", background = "CLL_MM_MCL")
)
}
\arguments{
\item{result_df}{TODO.}

\item{function_name}{TODO.}

\item{clobber_mode}{TODO.}

\item{get_existing}{TODO.}

\item{function_params}{TODO.}

\item{additional_details}{TODO.}
}
\value{
full path to the file that was written
}
\description{
TODO.
}
