% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tmb.R
\name{calculate_tmb}
\alias{calculate_tmb}
\title{Return TMB counts.}
\usage{
calculate_tmb(
  maf_data,
  regions_bed,
  projection = "grch37",
  subset_to_nonSyn = TRUE,
  vc_nonSyn = NULL
)
}
\arguments{
\item{maf_data}{Incoming data frame representing simple somatic
mutations in maf format. The minimal required columns are
Tumor_Sample_Barcode, Hugo_Symbol, Chromosome, Start_Position, End_Position,
Variant_Classification.}

\item{regions_bed}{Optionally, specify the data frame with sequenced panel.
Expected to be in bed format with first 3 columns indicating chromosome,
start, and end positions. This function is agnostic to the naming of these
columns. If there are any other columns present, they are ignored.}

\item{projection}{The genome build of projection for the data in the incoming
maf. Default is "grch37".}

\item{subset_to_nonSyn}{Logical argument to control whether the maf file
should be subset to non-synonymous mutations only. Defaults to TRUE (perform
subsetting).}

\item{vc_nonSyn}{Vector of annotations from Variant_Classification column
that should be considered non-synonymous.}
}
\value{
data frame
}
\description{
This function implements tumor mutation burden calculation.
TODO: add more details.
}
\examples{
# obtain maf data
library(GAMBLR.data)
maf1 <- get_ssm_by_samples()

calculate_tmb(maf1)
calculate_tmb(
    maf1,
    regions_bed = grch37_ashm_regions
)
#' calculate_tmb(
    maf1,
    regions_bed = grch37_ashm_regions,
    subset_to_nonSyn = FALSE
)

}
