% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config_value.R
\name{check_config_and_value}
\alias{check_config_and_value}
\title{Check Config and Value.}
\usage{
check_config_and_value(config_key, config_name)
}
\arguments{
\item{config_key}{A string of one or more nested keys to retrieve
from the config in the format "$key1$key2$keyN"}
}
\value{
A string with the value from the config key or nothing
(if config key is NULL or missing).
}
\description{
Check the existence of a specific config key.
The function will notify the user and end the program if no such key exists.
}
\details{
INTERNAL FUNCTION for checking the existence of a config value,
not meant for out-of-package usage.
}
\examples{
check_config_value(config::get("resources")$blacklist$template)

}
