% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_details.R
\name{check_file_details}
\alias{check_file_details}
\title{Check file details}
\usage{
check_file_details(relative_paths)
}
\arguments{
\item{relative_paths}{Vector of relative paths.}
}
\value{
Boolean
}
\description{
When relative path to a file is given, this function will
automatically generate it's full path using the project_base config value
and check for existence of the file at the full path. Can operate on vector
of several relative paths.
}
