% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-vars.R
\docType{data}
\name{cnames}
\alias{cnames}
\title{Required columns in bedpe.}
\format{
An object of class \code{character} of length 19.
}
\usage{
cnames
}
\description{
Global variable specifying what bedpe columns are absolutely
required.
}
\keyword{datasets}
