% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-vars.R
\docType{data}
\name{coding_class}
\alias{coding_class}
\title{Coding+Silent variants classification.}
\format{
An object of class \code{character} of length 12.
}
\usage{
coding_class
}
\description{
Same as coding_vc, but includes Silent variants to consider
them as coding.
}
\keyword{datasets}
