% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-vars.R
\docType{data}
\name{coding_vc}
\alias{coding_vc}
\title{Coding variants classification.}
\format{
An object of class \code{character} of length 11.
}
\usage{
coding_vc
}
\description{
Vector of variant classification values considered as coding.
}
\keyword{datasets}
