% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-vars.R
\docType{data}
\name{colour_aliases}
\alias{colour_aliases}
\title{Color code aliases.}
\format{
An object of class \code{list} of length 17.
}
\usage{
colour_aliases
}
\description{
A list of alias namings for color schemes.
}
\keyword{datasets}
