% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_coding_mutation_pattern.R
\name{compare_coding_mutation_pattern}
\alias{compare_coding_mutation_pattern}
\title{Compare pattern of coding mutations.}
\usage{
compare_coding_mutation_pattern(maf_df1, maf_df2, gene)
}
\arguments{
\item{maf_df1}{TODO.}

\item{maf_df2}{TODO.}

\item{gene}{TODO.}
}
\value{
list
}
\description{
TODO.
}
