% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_bed.R
\name{decorate_bed}
\alias{decorate_bed}
\title{Decorate a bed or bedpe file}
\usage{
decorate_bed(
  bed_df,
  colour_by = "CHROM_B",
  colour_mapping,
  new_column_name = "color"
)
}
\arguments{
\item{bed_df}{A bed or bedpe data frame}

\item{colour_by}{Specify the name of the column used to define the colours}

\item{colour_mapping}{Named vector specifying the colour for each possible value in the colour_by column}

\item{new_column_name}{The name to use for the colour column (default is 'color' for IGV compatability)}
}
\value{
data frame with one additional column containing colours for each row
}
\description{
Decorate a bed or bedpe file
}
\examples{

in_bed = get_combined_sv()
coloured_bed = decorate_bed(in_bed,
                            colour_mapping = get_gambl_colours("chromosome",
                                                              as_rgb_string = T))
}
