% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMBLR.helpers-package.R
\docType{package}
\name{GAMBLR.helpers-package}
\alias{GAMBLR.helpers}
\alias{GAMBLR.helpers-package}
\title{GAMBLR.helpers: Provide helper functions for GAMBLR-based packages}
\description{
This package is part of the Genomic Analysis of Mature B-cell Lymphomas (GAMBL) project developed by the Morin Lab.
}
\author{
\strong{Maintainer}: Vladimir Souza \email{vsouza@bcgsc.ca}

}
\keyword{internal}
