% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mutation_tally.R
\name{gene_mutation_tally}
\alias{gene_mutation_tally}
\title{Tally coding mutations load.}
\usage{
gene_mutation_tally(
  maf_df,
  these_samples_metadata,
  these_genes,
  grouping_variable = "cohort"
)
}
\arguments{
\item{maf_df}{Data frame of simple somatic mutations in maf format. Required
parameter.}

\item{these_samples_metadata}{Data frame with metadata. Must contain sample
identifiers in the \code{sample_id} column and column that will be used
to calulate the mutation load. All other columns are ignored. Required
parameter.}

\item{these_genes}{Vector of hugo symbols for genes to be considered for the
tallying of mutations. Required parameter.}

\item{grouping_variable}{Column in the metadata that will be used as grouping
variable. By default, the \code{cohort} is used.}
}
\value{
data frame
}
\description{
Calculate the number of coding mutations per grouping. Silent
variants are excluded from this calculation. The incoming maf may contain
non-coding variants, which will be excluded from tallying.
}
