% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template_wildcards.R
\name{get_template_wildcards}
\alias{get_template_wildcards}
\title{Get template wildcards.}
\usage{
get_template_wildcards(parent_key, template_key)
}
\arguments{
\item{parent_key}{TODO.}

\item{template_key}{TODO.}
}
\value{
list
}
\description{
TODO.
}
