% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unmatched_normals.R
\name{get_unmatched_normals}
\alias{get_unmatched_normals}
\title{Get unmatched normals.}
\usage{
get_unmatched_normals(seq_type_filter)
}
\arguments{
\item{seq_type_filter}{Seq type key from config for which to return the
unmatched normals for.}
}
\value{
data frame
}
\description{
Helper function to get the unmatched normals from the main
config.
}
