% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_wildcards.R
\name{grob_wildcards}
\alias{grob_wildcards}
\title{Glob wildcards.}
\usage{
grob_wildcards(wildcarded_string)
}
\arguments{
\item{wildcarded_string}{String containing wildcards inside the {} notations.}
}
\value{
vector
}
\description{
Helper function to extract wildcard values from a string.
}
