% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_metadata.R
\name{handle_metadata}
\alias{handle_metadata}
\title{Handle metadata from bundled packages.}
\usage{
handle_metadata(this_seq_type = "genome")
}
\arguments{
\item{this_seq_type}{The seq type of the samples to return with metadata.}
}
\value{
data frame
}
\description{
Will return the metadata from the GAMBLR.data package when it is
installed. Otherwise, will instruct user to call
GAMBLR.results::get_gambl_metadata().
}
\details{
INTERNAL FUNCTION for handling the metadata.
}
