% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_ssm_by_region.R
\name{handle_ssm_by_region}
\alias{handle_ssm_by_region}
\title{Handle SSM by Region}
\usage{
handle_ssm_by_region(this_maf, chromosome, qstart, qend, region = "")
}
\arguments{
\item{this_maf}{An already loaded MAF or MAF-like object to subset to regions of interest. This is a required paremter.}

\item{chromosome}{The chromosome you are restricting to (with or without a chr prefix).}

\item{qstart}{Query start coordinate of the range you are restricting to.}

\item{qend}{Query end coordinate of the range you are restricting to.}

\item{region}{Region formatted like chrX:1234-5678 instead of specifying chromosome, start and end separately.}
}
\value{
A MAF that has been subset to the regions specified.
}
\description{
Subset an incoming MAF to desired regions.
}
\details{
This helper function is used to subset an incoming MAF to specific region.
This function was developed to accommodate non-GSC-access users to use plotting functions that, 
previously internally called GABLR core functions (GSC access required).
Thsi function takes an already loaded MAF (`this_maf`) and subset this based on the regions provided,
either in region format (chrX:1234-5678) withh the `region` parameter.
Or, the suer can seperatly specify the regions of interest using the following parameters;
`chromosome`, `qstart`, and, `qend`.
}
\examples{
my_maf = GAMBLR.data::sample_data$grch37$maf
my_mutations = handle_ssm_by_region(this_maf = my_maf, region = "chr8:128,723,128-128,774,067")

}
