% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_chr_sizes_from_bundled_data.R
\name{infer_chr_sizes_from_bundled_data}
\alias{infer_chr_sizes_from_bundled_data}
\title{Infer chromosome sizes using the bundled data}
\usage{
infer_chr_sizes_from_bundled_data(projection = "grch37")
}
\arguments{
\item{projection}{Specify which genome build to use. One of "grch37" (default) or "hg38".}
}
\value{
A data frame with columns \code{chromosome} and \code{size.}
}
\description{
Infer chromosome sizes using the bundled data
}
\examples{
infer_chr_sizes_from_bundled_data(projection = "hg38")

}
