% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-vars.R
\docType{data}
\name{maf_header}
\alias{maf_header}
\title{Column names in maf.}
\format{
An object of class \code{numeric} of length 116.
}
\usage{
maf_header
}
\description{
A vector of column names and their order in standard maf file.
}
\keyword{datasets}
