% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-vars.R
\docType{data}
\name{rainfall_conv}
\alias{rainfall_conv}
\title{Reverse mappings.}
\format{
An object of class \code{character} of length 13.
}
\usage{
rainfall_conv
}
\description{
A vector of substitutions mappings for the rainfall plots.
}
\keyword{datasets}
