% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_bed.R
\name{reciprocate_bedpe}
\alias{reciprocate_bedpe}
\title{Add reciprocal events to a bedpe file}
\usage{
reciprocate_bedpe(bed_df, shift_by = 30)
}
\arguments{
\item{bed_df}{}

\item{shift_by}{The positions of START_A and END_A will be increased by this value (default 30). When combining this with decorate_bed, this offset ensures IGV shows the intended colour instead of the colour of the reciprocal event.}
}
\value{
data frame with each row duplicated and the contents of CHROM_A, START_A etc swapped with CHROM_B, START_B
}
\description{
Add reciprocal events to a bedpe file
}
\examples{


recip_bed = reciprocate_bedpe(in_bed)
coloured_bed = decorate_bed(recip_bed,
                            is_bedpe=T,
                            colour_mapping = get_gambl_colours("chromosome",
                                                                as_rgb_string = T))
                            
}
