% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-vars.R
\docType{data}
\name{required_cols}
\alias{required_cols}
\title{Required columns in metadata.}
\format{
An object of class \code{character} of length 7.
}
\usage{
required_cols
}
\description{
Global variable specifying what metadata columns are absolutely
required.
}
\keyword{datasets}
