% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_cnstates.R
\name{subset_cnstates}
\alias{subset_cnstates}
\title{Subset CN States. DEPRECATED}
\usage{
subset_cnstates(cn_segments, include_2 = FALSE, samplen)
}
\arguments{
\item{cn_segments}{DF with copy number segments, usually retrieved from get_sample_cn_segments.}

\item{include_2}{Optional parameter for including or omit CN state == 2. Default is FALSE.}

\item{samplen}{Numeric value that annotates the sample order.}
}
\value{
Nothing
}
\description{
Get the available CN states in the incoming data frame.
}
\details{
For sub-setting copy number information based on segments available in cn data
}
\examples{
\dontrun{
meta = get_gambl_metadata() \%>\% dplyr::filter(sample_id \%in\% c("02-13135T","SU-DHL-4"))
cn_states = get_cn_segments(these_samples_metadata=meta)

subsetted = subset_cnstates(cn_segments = cn_states,
                samplen = 1)
dim(subsetted)
}
}
