% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-vars.R
\docType{data}
\name{vc_nonSynonymous}
\alias{vc_nonSynonymous}
\title{Non-synonymous variants classification.}
\format{
An object of class \code{character} of length 9.
}
\usage{
vc_nonSynonymous
}
\description{
Vector of variant classification values considered as
non-synonymous.
}
\keyword{datasets}
