% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_onco_matrix.R
\name{create_onco_matrix}
\alias{create_onco_matrix}
\title{Create onco matrix from maf data.}
\usage{
create_onco_matrix(maf_df, genes, add_missing = TRUE, subset_to_coding = FALSE)
}
\arguments{
\item{maf_df}{Data frame with maf data. Required parameter. The minimal required columns are Tumor_Sample_Barcode, Hugo_Symbol, Variant_Classification, Start_Position, and End_Position.}

\item{genes}{List of genes to return in the resulting matrix. When not provided, matrix is generated for each gene present in the input maf data.}

\item{add_missing}{When list of genes is provided and some of the specified genes are not mutated in the provided maf data, they will still be added to the matrix with 0\% mutation frequency. Default TRUE.}

\item{subset_to_coding}{Will conditionally subset to only coding variants. Default is FALSE (no subsetting).}
}
\value{
matrix
}
\description{
This function implements generation of ComplexHeatmap-compatible
matrix using dplyr. The output of this function is identical to the implementation
in maftools, hovewer the main advantage of this version is tidy syntax and
avoidance of data.table dependency since the latter package cannot handle
large tables required for processing of large multi-sample maf files.
}
\examples{
library(GAMBLR.data)
maf <- get_coding_ssm()
onco_matrix <- create_onco_matrix(maf_df = maf)

}
