% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_hotspots.R
\name{review_hotspots}
\alias{review_hotspots}
\title{Review Hotspots.}
\usage{
review_hotspots(
  annotated_maf,
  genes_of_interest = c("FOXO1", "MYD88", "CREBBP", "NOTCH1", "NOTCH2", "CD79B", "EZH2"),
  genome_build,
  custom_coordinates
)
}
\arguments{
\item{annotated_maf}{A data frame in MAF format that has hotspots annotated using the function annotate_hotspots().}

\item{genes_of_interest}{A vector of genes for hotspot review. Currently only FOXO1, MYD88, CREBBP, NOTCH1, NOTCH2, CD79B and EZH2 are supported.}

\item{genome_build}{Reference genome build for the coordinates in the MAF file. The default is grch37 genome build.}

\item{custom_coordinates}{A data frame with custom coordinates for the hot spots.
All mutations in any of the regions specified in the data frame will be marked as hot spots.
The data frame must have the following columns: "Hugo_Symbol", "chrom", "start", and "end".}
}
\value{
The same data frame (as given to the \code{annotated_maf} parameter) with the reviewed column "hot_spot".
}
\description{
Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations.
}
\details{
This function takes an annotated MAF (with annotate_hotspots) and updates an existing column, "hot_spot", in the same data frame.
Genes for hotspot review are supplied with the \code{genes_of_interest} parameter.
Currently only a few sets of genes are supported, see parameter description for more information and limitations.
The desired genome build can be specified with \code{genome_build} parameter. Should be the same as the incoming MAF.
}
\examples{
hot_ssms = review_hotspots(annotate_hotspots(get_coding_ssm(this_seq_type = "genome")),
                           genes_of_interest = c("CREBBP"))

}
