% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_ensg_ids.R
\name{translate_ensg_ids}
\alias{translate_ensg_ids}
\title{Translate ENSG ids.}
\usage{
translate_ensg_ids(
  df,
  ensg_column_in = "ensembl_gene_id",
  symbol_column_out = "hgnc_symbol",
  ids_in_rownames = FALSE
)
}
\arguments{
\item{df}{an incoming data frame with ENSG ids either in column or rownames.}

\item{ensg_column_in}{string specifying column name where to find the ENSG
ids. Only used when \code{ids_in_rownames} is set to FALSE. Default is
"gene_id".}

\item{symbol_column_out}{string specifying column name where to output the
ENSG ids. Only used when \code{ids_in_rownames} is set to FALSE. Default is
"gene_id".}

\item{ids_in_rownames}{boolean specifying whether ENSG ids are in rownames or
in a designated column. Default is FALSE (the ENSG ids are a column).}
}
\value{
df
}
\description{
Translate gene ids in ENSG formats (with or without trailing
.\if{html}{\out{<number>}}) to a human-readable format of Hugo symbols. Note that there
are more ENSG ids then there are hugo symbols: 20,075 of ENSG ids don't
have associated Hugo symbol. Those are novel transcripts, lncRNAS,
pseudogenes etc, and they will be dropped from the resulting output of
this function.
}
\examples{
test_this <- GAMBLR.data::gencode_to_symbol \%>\%
    select(1) \%>\%
    head(10) \%>\%
    mutate(test = "test")

translate_ensg_ids(
    df = test_this \%>\%
        column_to_rownames("ensembl_gene_id"),
    ids_in_rownames = TRUE
)

translate_ensg_ids(
    df = test_this
)

translate_ensg_ids(
    df = test_this \%>\%
        select(
            test_col = ensembl_gene_id,
            everything()
        ),
    ensg_column_in = "test_col"
)

translate_ensg_ids(
    df = test_this \%>\%
        select(
            test_col = ensembl_gene_id,
            everything()
        ),
    ensg_column_in = "test_col",
    symbol_column_out = "this, Ha!"
)


}
